feat: enhance WorkerService methods with error handling for existing workers

+++ Back-end/src/Modules/worker/worker.service.ts
@@ -4,34 +4,58 @@ import { UpdateWorkerDto } from './dto/update-worker.dto';
 import { PrismaService } from '../../prisma.service';
 import { Worker } from '@prisma/client';
 
-	@Injectable()
-	export class WorkerService { 
-  constructor(private Prisma: PrismaService) {}
+@Injectable()
+export class WorkerService {
+	constructor(private Prisma: PrismaService) { }
 
-  async create(createWorkerDto: CreateWorkerDto): Promise<Worker> {
-    return this.Prisma.worker.create({ data: createWorkerDto });
-  }
+	async create(createWorkerDto: CreateWorkerDto): Promise<Worker> {
+		const existingWorkerUserId = await this.Prisma.worker.findUnique({
+			where: { userId: createWorkerDto.userId }
+		});
+		if (existingWorkerUserId) {
+			throw new Error('Worker with this userId already exists');
+		}
 
-  findAll() {
-		return this.Prisma.worker.findMany()
-  }
+		return this.Prisma.worker.create({ data: createWorkerDto });
+	}
 
-  findOne(id: number) {
-    return this.Prisma.worker.findUnique({
+	findAll() {
+		return this.Prisma.worker.findMany();
+	}
+
+	async findOne(id: number) {
+		const worker = await this.Prisma.worker.findUnique({
 			where: { id }
 		});
-  }
+		if (!worker) {
+			throw new Error(`Worker with this id ${id} not found`);
+		}
+		return worker;
+	}
 
-  update(id: number, updateWorkerDto: UpdateWorkerDto) {
-    return this.Prisma.worker.update({
+	async update(id: number, updateWorkerDto: UpdateWorkerDto) {
+		const updateworker = await this.Prisma.worker.findUnique({
+			where: { id }
+		});
+		if (!updateworker) {
+			throw new Error(`Worker with this id ${id} not found`);
+		}
+		const updatedWorker = await this.Prisma.worker.update({
 			where: { id },
 			data: updateWorkerDto
-		})
-  }
+		});
+		return updatedWorker;
+	}
 
-  remove(id: number) {
-    return this.Prisma.worker.delete({
-			where : { id }
-		})
-  }
+	async remove(id: number) {
+		const deleteworker = await this.Prisma.worker.findUnique({
+			where: { id }
+		});
+		if (!deleteworker) {
+			throw new Error(`Worker with this id ${id} not found`);
+		}
+		return this.Prisma.worker.delete({
+			where: { id }
+		});
+	}
 }
